LPS-53627 extract conversions functionality to its own module
