+++ modules/apps/journal/journal-content-asset-addon-entries/journal-content-conversions-asset-addon-entry/src/com/liferay/journal/content/asset/addon/entry/conversions/BaseConvertionUserToolAssetAddonEntry.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.journal.content.web.entries;
+package com.liferay.journal.content.asset.addon.entry.conversions;
 
 import com.liferay.journal.content.web.util.UserToolAssetAddonEntry;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.servlet.taglib.ui.BaseAssetAddonEntry;
+import com.liferay.portal.kernel.servlet.taglib.ui.BaseJSPAssetAddonEntry;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.PrefsPropsUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
@@ -24,16 +24,26 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.util.PropsValues;
 import com.liferay.portlet.documentlibrary.util.DocumentConversionUtil;
 
+import java.io.IOException;
+
 import java.util.Locale;
 import java.util.ResourceBundle;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author Julio Camarero
  */
 public abstract class BaseConvertionUserToolAssetAddonEntry
-	extends BaseAssetAddonEntry implements UserToolAssetAddonEntry {
+	extends BaseJSPAssetAddonEntry implements UserToolAssetAddonEntry {
 
 	public abstract String getExtension();
+
+	@Override
+	public String getJSPPath() {
+		return _JSP_PATH;
+	}
 
 	@Override
 	public String getLabel(Locale locale) {
@@ -43,6 +53,16 @@ public abstract class BaseConvertionUserToolAssetAddonEntry
 		return LanguageUtil.format(
 			resourceBundle, "download-as-x",
 			StringUtil.toUpperCase(getExtension()));
+	}
+
+	@Override
+	public void include(
+			HttpServletRequest request, HttpServletResponse response)
+		throws IOException {
+
+		request.setAttribute("extension", getExtension());
+
+		super.include(request, response);
 	}
 
 	@Override
@@ -63,5 +83,8 @@ public abstract class BaseConvertionUserToolAssetAddonEntry
 
 		return super.isEnabled();
 	}
+
+	private static final String _JSP_PATH =
+		"/META-INF/resources/conversions.jsp";
 
 }
