+++ modules/apps/journal/journal-content-asset-addon-entries/journal-content-conversions-asset-addon-entry/src/com/liferay/journal/content/asset/addon/entry/conversions/PDFConvertionUserToolAssetAddonEntry.java
@@ -12,11 +12,14 @@
  * details.
  */
 
-package com.liferay.journal.content.web.entries;
+package com.liferay.journal.content.asset.addon.entry.conversions;
 
 import com.liferay.journal.content.web.util.UserToolAssetAddonEntry;
 
+import javax.servlet.ServletContext;
+
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Julio Camarero
@@ -24,28 +27,36 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true, service = UserToolAssetAddonEntry.class
 )
-public class TXTConvertionUserToolAssetAddonEntry
+public class PDFConvertionUserToolAssetAddonEntry
 	extends BaseConvertionUserToolAssetAddonEntry
 	implements UserToolAssetAddonEntry {
 
 	@Override
 	public String getExtension() {
-		return "txt";
+		return "pdf";
 	}
 
 	@Override
 	public String getIcon() {
-		return "file-alt";
+		return "file-text";
 	}
 
 	@Override
 	public String getKey() {
-		return "enableTXT";
+		return "enablePDF";
 	}
 
 	@Override
 	public Double getWeight() {
-		return 6.0;
+		return 3.0;
+	}
+
+	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.journal.content.asset.addon.entry.conversions)"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
 	}
 
 }
