+++ modules/apps/journal/journal-content-asset-addon-entries/journal-content-conversions-asset-addon-entry/src/content/Language.properties
@@ -1,0 +1,2 @@
+download-as-x=Download as {0}
+x-download-x-as-x=<span class="{0}">Download {1} as</span> {2}
