+++ modules/apps/journal/journal-content-web/src/content/Language.properties
@@ -2,7 +2,6 @@ select-user-tools=Select User Tools
 select-content-metadata=Select Content Metadata
 user-tools=User Tools
 displaying-content=Displaying Content
-download-as-x=Download as {0}
 javax.portlet.title.com_liferay_journal_content_web_portlet_JournalContentPortlet=Web Content Display
 please-select-a-web-content-from-the-list-below=Please select web content from the list below.
 select-existing-web-content-or-add-some-web-content-to-be-displayed-in-this-portlet=Select existing web content or add some web content to be displayed in this portlet.
